# src/internal/docs/swagger.yaml
basePath: /
definitions:
  annot_handler.RequestID:
    properties:
      id:
        type: integer
    type: object
  annot_handler.RequestUpdate:
    properties:
      class_label:
        type: integer
      error_bb:
        items:
          type: number
        type: array
      id:
        type: integer
      is_valid:
        type: boolean
      type_label:
        type: integer
    type: object
  annot_handler.ResponseGetAnnot:
    properties:
      class_label:
        type: integer
      creator_id:
        type: integer
      error:
        type: string
      error_bb:
        items:
          type: number
        type: array
      id:
        type: integer
      isValid:
        type: boolean
      page_data:
        items:
          type: integer
        type: array
      status:
        type: string
      type_label:
        type: integer
      was_checked:
        type: boolean
    type: object
...
  annot_type_handler.RequestAnnotType:
    properties:
      class_name:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
  annot_type_handler.RequestID:
    properties:
      id:
        type: integer
    type: object
  annot_type_handler.RequestIDs:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ggwpezsmth
    name: Andrew
    url: http://www.swagger.io/support
  description: An App for storing, getting annots and auto checking documents.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Annoter
  version: "1.0"
paths:
  /annot/add:
    post:
      consumes:
      - multipart/form-data
      description: Adds an annotation to a specific document
      parameters:
      - description: PNG image to add
        in: formData
        name: annotFile
        required: true
        type: file
      - in: formData
        name: class_label
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: number
        name: error_bb
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add an annotation
      tags:
      - Annotation
  /annot/check:
    post:
      consumes:
      - application/json
      description: Updating chosed markup and marking it as checked as well as setting
        creator_id to the current logged user
      parameters:
      - description: data to fix broken markup
        in: body
        name: NewMarkupParams
        required: true
        schema:
          $ref: '#/definitions/annot_handler.RequestUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Modifies markup and marks it as checked
      tags:
      - Annotation
  /annot/creatorID:
    get:
      consumes:
      - application/json
      description: Getting all anottattions from a database, which were created by
        currently logged user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Getting all anotattions created by a user
      tags:
      - Annotation
  /annot/get:
    post:
      consumes:
      - application/json
      description: Get the specific annotation by ID
      parameters:
      - description: Annot ID
        in: body
        name: AnnotID
        required: true
        schema:
          $ref: '#/definitions/annot_handler.RequestID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/annot_handler.ResponseGetAnnot'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a specific annotation
      tags:
      - Annotation
  /annot/getsAll:
    get:
      consumes:
      - application/json
      description: Getting all anottattions from a database, works only when there
        are not a lot of annotattions has no paging
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Getting all anottattions from a database
      tags:
      - Annotation
  /annotType/add:
    post:
      consumes:
      - application/json
      description: Create and save the new anotattion type, as created by signed in
        user
      parameters:
      - description: data for inserting new annotType
        in: body
        name: NewAnnotTypeParams
        required: true
        schema:
          $ref: '#/definitions/annot_type_handler.RequestAnnotType'
      produces:
      - application/json
      responses:
        "200":
          description: Annotation type not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add new anotattion type
      tags:
      - Annotation types
  /annotType/creatorID:
    get:
      consumes:
      - application/json
      description: Get all anotattions which were created by specific user
      produces:
      - application/json
      responses:
        "200":
          description: Annotation type not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a annotation type of a signed in user
      tags:
      - Annotation types
  /annotType/delete:
    delete:
      consumes:
      - application/json
      description: Delete an anotattion by specific ID
      parameters:
      - description: ID for deleting an annot
        in: body
        name: AnnotTypeID
        required: true
        schema:
          $ref: '#/definitions/annot_type_handler.RequestID'
      produces:
      - application/json
      responses:
        "200":
          description: Annotation type not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete AnnotType By userID
      tags:
      - Annotation types
  /annotType/gets:
    post:
      consumes:
      - application/json
      description: Extracts numerous types for a set of IDs
      parameters:
      - description: data for getting numerous anotattions
        in: body
        name: GetAnnotTypesIDs
        required: true
        schema:
          $ref: '#/definitions/annot_type_handler.RequestIDs'
      produces:
      - application/json
      responses:
        "200":
          description: Annotation not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get numerous types by numerous IDs
      tags:
      - Annotation types
  /annotType/getsAll:
    get:
      consumes:
      - application/json
      description: get all available annot Types
      produces:
      - application/json
      responses:
        "200":
          description: Annotation type not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Getting all available annot types
      tags:
      - Annotation types
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"